requirements.txt
selenium==3.141.0
black==19.10b0
python-twitter==3.5
pytest==5.3.3
ipython==7.11.1